requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.11.0
python-dotenv>=1.0.1
pytest>=8.3.0
httpx>=0.28.0
PyJWT>=2.9.0
